Modified the machine to make and destroy references to environments.
When using mach->env, it destroys the reference before over-writing the
pointer but I'm not sure I'm going to stick with this idea.  It means that
when you change mach->env in library functions (which you shouldn't really
do anyways) you have to remember to destroy the reference, even if you are
using sdrl_extend_environment.  Maybe you should change how references are
managed in bindings.c to make it easier.