Added the dblock and lblock types and change the expr type.
The dblock type is now what the old expr type was (a dynamically scoped AST
type).  The expr type is now a non-scoped pure AST type (which currently
ignores the args passed to it).  The lblock type is a lexically scoped AST
type which stores the current environment when a value is created and extends
that when it evaluates itself instead of using the current environment at the
time of evaluation.  It's still a little kludgy in it's implementation.  The
code form creates the new expr type and 2 new forms where added to create
dblock and lblock types (the dblock and lblock forms).  This means that
existing code must be rewritten.