Completely change the value system to be "Object-based"
The value struct itself does not contain any data and is used as an inherited
object by data-containing structs.  The whole thing is allocated at once
according to the size of the type (which is not fully implemented yet).  This
allows both sdrl_environment and sdrl_expr to be castable to sdrl_value.  All
reference-countable objects will inherit from sdrl_value.  The type functions
now deal directly with the value and not a generic pointer to data.  This
should make things a lot more flexable.  The type system itself hasn't really
been changed.  I suspect the fact that we must lookup a type all the time
causes massive performance penalties.  This code has not been cleaned up and is
not consistent in all places.  It should compile though.